requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
blis==0.7.6
brotlipy==0.7.0
cachetools==5.0.0
catalogue==2.0.6
certifi==2021.10.8
click==8.0.4
cycler==0.11.0
cymem==2.0.6
datasets==2.0.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
detoxify==0.4.0
dill==0.3.4
docker-pycreds==0.4.0
entrypoints==0.4
executing==0.8.3
filelock==3.6.0
fonttools==4.31.2
frozenlist==1.3.0
fsspec==2022.2.0
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
httplib2==0.20.4
huggingface-hub==0.4.0
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.0
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
multiprocess==0.70.12.2
murmurhash==1.0.6
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.10
numpy==1.22.2
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
preshed==3.0.6
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pydantic==1.8.2
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
regex==2022.3.2
responses==0.18.0
rsa==4.8
sacremoses==0.0.47
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
sentencepiece==0.1.96
sentry-sdk==1.5.6
setproctitle==1.2.2
shortuuid==1.0.8
smart-open==5.2.1
smmap==5.0.0
soupsieve==2.3.1
spacy==3.2.3
spacy-legacy==3.0.9
spacy-loggers==1.0.1
scrubadub==2.0.0
srsly==2.4.2
stack-data==0.2.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
thinc==8.0.15
tokenizers==0.11.6
torch==1.10.2
tornado==6.1
traitlets==5.1.1
transformers==4.17.0
typer==0.4.0
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.7
wandb==0.12.11
wasabi==0.9.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
xxhash==3.0.0
yarl==1.7.2
yaspin==2.1.0
zstandard==0.17.0
pyap==0.3.1
pycodestyle==2.8.0
scrubadub_spacy
openai==0.26.5
